requirements.txt
# Core Data Libraries
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
matplotlib==3.10.0
seaborn==0.13.2

# PyTorch Ecosystem (CPU-only)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Computer Vision
opencv-python-headless==4.10.0.84
pillow==10.4.0
imageio==2.37.0
imageio-ffmpeg==0.6.0

# NLP & Transformers
transformers>=4.30.0,<4.50.0
sentence-transformers==4.1.0
nltk==3.9.1
regex==2024.11.6

# LangChain & Vector Stores
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
faiss-cpu==1.11.0

# OCR & Document Processing
easyocr==1.7.2
pytesseract==0.3.13
PyPDF2==3.0.1
python-docx==1.1.2

# Llama Cpp Python for embeddings
llama_cpp_python==0.3.9

# Geospatial
folium==0.19.5
geopandas==1.0.1
geopy==2.4.1
h3==4.2.2
keplergl==0.3.2

# Object Detection (YOLOv5 from GitHub)
-e ./yolov5_custom

# Hugging Face Hub
huggingface-hub>=0.26.0,<1.0

# MongoDB & Networking
pymongo==4.13.0
dnspython==2.7.0

# Streamlit & UI
streamlit==1.44.1
streamlit_chat==0.0.2.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6

# Earth Engine & Satellite Analysis
earthengine-api==1.5.11
geemap==0.35.3
rasterio==1.3.10
sentinelsat==1.1.1

# Web Scraping & APIs
requests==2.32.3
beautifulsoup4==4.13.3
praw==7.8.1
tldextract==5.1.1

# TensorFlow Ecosystem (CPU-compatible)
tensorflow==2.14.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0

# Utility
tqdm==4.67.1
joblib==1.3.2
python-dotenv==1.0.1

# OpenAI + Pinecone
openai==1.78.1
pinecone-client==2.2.1

# Security
cryptography==44.0.2

# Downloader
gdown==5.1.0
